style: remove redundant type casts

+++ src/net/sourceforge/plantuml/AnimatedGifEncoder.java
@@ -594,11 +594,11 @@ class NeuQuant {
 	/* defs for freq and bias */
 	protected static final int intbiasshift = 16; /* bias for fractions */
 
-	protected static final int intbias = (((int) 1) << intbiasshift);
+	protected static final int intbias = (1 << intbiasshift);
 
 	protected static final int gammashift = 10; /* gamma = 1024 */
 
-	protected static final int gamma = (((int) 1) << gammashift);
+	protected static final int gamma = (1 << gammashift);
 
 	protected static final int betashift = 10;
 
@@ -614,7 +614,7 @@ class NeuQuant {
 
 	protected static final int radiusbiasshift = 6; /* at 32.0 biased by 6 bits */
 
-	protected static final int radiusbias = (((int) 1) << radiusbiasshift);
+	protected static final int radiusbias = (1 << radiusbiasshift);
 
 	protected static final int initradius = (initrad * radiusbias); /*
 																	 * and
@@ -627,18 +627,18 @@ class NeuQuant {
 	/* defs for decreasing alpha factor */
 	protected static final int alphabiasshift = 10; /* alpha starts at 1.0 */
 
-	protected static final int initalpha = (((int) 1) << alphabiasshift);
+	protected static final int initalpha = (1 << alphabiasshift);
 
 	protected int alphadec; /* biased by 10 bits */
 
 	/* radbias and alpharadbias used for radpower calculation */
 	protected static final int radbiasshift = 8;
 
-	protected static final int radbias = (((int) 1) << radbiasshift);
+	protected static final int radbias = (1 << radbiasshift);
 
 	protected static final int alpharadbshift = (alphabiasshift + radbiasshift);
 
-	protected static final int alpharadbias = (((int) 1) << alpharadbshift);
+	protected static final int alpharadbias = (1 << alpharadbshift);
 
 	/*
 	 * Types and Global Variables --------------------------
@@ -1004,7 +1004,7 @@ class NeuQuant {
 		int bestpos, bestbiaspos, bestd, bestbiasd;
 		int[] n;
 
-		bestd = ~(((int) 1) << 31);
+		bestd = ~(1 << 31);
 		bestbiasd = bestd;
 		bestpos = -1;
 		bestbiaspos = bestpos;
